fix(global): MSVC(C4018)
signed/unsigned mismatch

+++ src/hooks/misc/start_matchmaking_find_sessions.cpp
@@ -31,7 +31,7 @@ namespace big
 							results[*num_sessions_found] = g_matchmaking_service->get_found_sessions()[i].info;
 							(*num_sessions_found)++;
 
-							if (max_sessions <= *num_sessions_found)
+							if (max_sessions <= (unsigned int)(*num_sessions_found))
 								break;
 						}
 					}
